Change cleanup containers logic.

See merge request olp/neutral-server/neutral-server-all-automation!204

+++ src/main/java/com/here/platform/cm/steps/remove/AllRemoveExtension.java
@@ -46,11 +46,12 @@ public class AllRemoveExtension implements AfterEachCallback {
         DataForRemoveCollector
                 .getAllArtificialPolicy(testId)
                 .forEach(policy -> new AaaCall().removeResourcePermission(policy.getKey(), policy.getValue()));
-        //TODO: remove after bam remove policy fix
-        DataForRemoveCollector
-                .getNsContainers(testId)
-                .forEach(container -> new AaaCall().deletePolicyForContainer(container));
-
+        if (!"prod".equalsIgnoreCase(System.getProperty("env"))) {
+            //TODO: remove after bam remove policy fix
+            DataForRemoveCollector
+                    .getNsContainers(testId)
+                    .forEach(container -> new AaaCall().deletePolicyForContainer(container));
+        }
         DataForRemoveCollector
                 .getNsContainers(testId)
                 .forEach(Steps::removeRegularContainer);
